move shared parameters to its own file
