+++ src/sc_keeper/api.py
@@ -1,17 +1,15 @@
 import logging
 from contextlib import asynccontextmanager
-from enum import Enum, StrEnum
 from importlib.metadata import version
 from os import environ
 from textwrap import dedent
-from types import SimpleNamespace
-from typing import Annotated, List, Literal, Optional
+from typing import Annotated, List, Literal
 
-from fastapi import Depends, FastAPI, HTTPException, Path, Query, Request, Response
+from fastapi import Depends, FastAPI, HTTPException, Path, Request, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
 from sc_crawler.table_bases import ServerBase
-from sc_crawler.table_fields import Allocation, CpuArchitecture, Status, StorageType
+from sc_crawler.table_fields import Status
 from sc_crawler.tables import (
     Benchmark,
     BenchmarkScore,
@@ -29,6 +27,7 @@ from sqlalchemy.exc import NoResultFound
 from sqlalchemy.orm import contains_eager
 from sqlmodel import Session, and_, func, not_, or_, select
 
+from . import parameters as options
 from .ai import openai_extract_filters
 from .currency import CurrencyConverter
 from .database import session
@@ -36,17 +35,12 @@ from .logger import LogMiddleware, get_request_id
 from .lookups import min_server_price
 from .query import max_score_per_server
 from .references import (
-    Countries,
-    Vendors,
-    Regions,
-    ComplianceFrameworks,
     ServerTableMetaData,
     ServerPKs,
     ServerPKsWithPrices,
     RegionPKs,
     ServerPriceWithPKs,
     OrderDir,
-    FilterCategories,
 )
 
 package_versions = {
@@ -223,221 +217,6 @@ app.add_middleware(
 # aggressive compression
 app.add_middleware(GZipMiddleware, minimum_size=100)
 
-
-# ##############################################################################
-# Shared parameters
-
-options = SimpleNamespace(
-    vendor=Annotated[
-        Optional[List[Vendors]],
-        Query(
-            title="Vendor id",
-            description="Identifier of the cloud provider vendor.",
-            json_schema_extra={
-                "category_id": FilterCategories.VENDOR,
-                "enum": [m.value for m in Vendors],
-            },
-        ),
-    ],
-    partial_name_or_id=Annotated[
-        Optional[str],
-        Query(
-            title="Partial name or id",
-            description="Freetext, case-insensitive search on the server_id, name, api_reference or display_name.",
-            json_schema_extra={
-                "category_id": FilterCategories.BASIC,
-            },
-        ),
-    ],
-    vcpus_min=Annotated[
-        int,
-        Query(
-            title="Minimum vCPUs",
-            description="Minimum number of virtual CPUs.",
-            ge=1,
-            le=256,
-            json_schema_extra={
-                "category_id": FilterCategories.PROCESSOR,
-                "unit": "vCPUs",
-            },
-        ),
-    ],
-    architecture=Annotated[
-        Optional[List[CpuArchitecture]],
-        Query(
-            title="Processor architecture",
-            description="Processor architecture.",
-            json_schema_extra={
-                "category_id": FilterCategories.PROCESSOR,
-                "enum": [e.value for e in CpuArchitecture],
-            },
-        ),
-    ],
-    memory_min=Annotated[
-        Optional[float],
-        Query(
-            title="Minimum memory",
-            description="Minimum amount of memory in GBs.",
-            json_schema_extra={
-                "category_id": FilterCategories.MEMORY,
-                "unit": "GB",
-                "step": 0.1,
-            },
-        ),
-    ],
-    price_max=Annotated[
-        Optional[float],
-        Query(
-            title="Maximum price",
-            description="Maximum price (USD/hr).",
-            json_schema_extra={
-                "category_id": FilterCategories.PRICE,
-                "step": 0.0001,
-            },
-        ),
-    ],
-    only_active=Annotated[
-        Optional[bool],
-        Query(
-            title="Active only",
-            description="Filter for active servers only.",
-            json_schema_extra={"category_id": FilterCategories.BASIC},
-        ),
-    ],
-    green_energy=Annotated[
-        Optional[bool],
-        Query(
-            title="Green energy",
-            description="Filter for regions with kow CO2 emission only.",
-            json_schema_extra={"category_id": FilterCategories.REGION},
-        ),
-    ],
-    allocation=Annotated[
-        Optional[Allocation],
-        Query(
-            title="Allocation",
-            description="Server allocation method.",
-            json_schema_extra={
-                "enum": [m.value for m in Allocation],
-            },
-        ),
-    ],
-    regions=Annotated[
-        Optional[List[Regions]],
-        Query(
-            title="region id",
-            description="Identifier of the region.",
-            json_schema_extra={
-                "category_id": FilterCategories.REGION,
-                "enum": [m.value for m in Regions],
-            },
-        ),
-    ],
-    compliance_framework=Annotated[
-        Optional[List[ComplianceFrameworks]],
-        Query(
-            title="Compliance Framework id",
-            description="Compliance framework implemented at the vendor.",
-            json_schema_extra={
-                "category_id": FilterCategories.VENDOR,
-                "enum": [m.value for m in ComplianceFrameworks],
-            },
-        ),
-    ],
-    storage_size=Annotated[
-        Optional[float],
-        Query(
-            title="Storage Size",
-            description="Minimum amount of storage (GBs) attached to the server.",
-            json_schema_extra={
-                "category_id": FilterCategories.STORAGE,
-                "step": 0.1,
-                "unit": "GB",
-            },
-        ),
-    ],
-    storage_type=Annotated[
-        Optional[List[StorageType]],
-        Query(
-            title="Storage Type",
-            description="Type of the storage attached to the server.",
-            json_schema_extra={
-                "category_id": FilterCategories.STORAGE,
-                "enum": [e.value for e in StorageType],
-            },
-        ),
-    ],
-    countries=Annotated[
-        Optional[List[str]],
-        Query(
-            title="Countries",
-            description="Filter for regions in the provided list of countries.",
-            json_schema_extra={
-                "category_id": FilterCategories.REGION,
-                "enum": [e.value for e in Countries],
-            },
-        ),
-    ],
-    gpu_min=Annotated[
-        Optional[int],
-        Query(
-            title="GPU count",
-            description="Minimum number of GPUs.",
-            json_schema_extra={
-                "category_id": FilterCategories.GPU,
-                "unit": "GPUs",
-            },
-        ),
-    ],
-    gpu_memory_min=Annotated[
-        Optional[float],
-        Query(
-            title="Minimum GPU memory",
-            description="Minimum amount of GPU memory (GB) in each GPU.",
-            json_schema_extra={
-                "category_id": FilterCategories.GPU,
-                "unit": "GB",
-                "step": 0.1,
-            },
-        ),
-    ],
-    gpu_memory_total=Annotated[
-        Optional[float],
-        Query(
-            title="Total GPU memory",
-            description="Minimum amount of total GPU memory (GBs) in all GPUs.",
-            json_schema_extra={
-                "category_id": FilterCategories.GPU,
-                "unit": "GB",
-                "step": 0.1,
-            },
-        ),
-    ],
-    benchmark_score_stressng_cpu_min=Annotated[
-        Optional[float],
-        Query(
-            title="SCore",
-            description="Minimum stress-ng CPU workload score.",
-            json_schema_extra={
-                "category_id": FilterCategories.PROCESSOR,
-            },
-        ),
-    ],
-    limit=Annotated[
-        int, Query(description="Maximum number of results. Set to -1 for unlimited.")
-    ],
-    limit250=Annotated[int, Query(description="Maximum number of results.", le=250)],
-    page=Annotated[Optional[int], Query(description="Page number.")],
-    order_by=Annotated[str, Query(description="Order by column.")],
-    order_dir=Annotated[OrderDir, Query(description="Order direction.")],
-    currency=Annotated[Optional[str], Query(description="Currency used for prices.")],
-    add_total_count_header=Annotated[
-        bool,
-        Query(
-            description="Add the X-Total-Count header to the response with the overall number of items (without paging). Note that it might reduce response times."
-        ),
-    ],
-)
 
 # ##############################################################################
 # API endpoints
